refactor: split system status api clients
