+++ src/ui/main-axios.ts
@@ -92,7 +92,7 @@ export type ServerMetrics = {
   lastChecked: string;
 };
 
-interface AuthResponse {
+export interface AuthResponse {
   success?: boolean;
   is_admin?: boolean;
   username?: string;
@@ -258,7 +258,7 @@ export function getCookie(name: string): string | undefined {
 let userWasAuthenticated = false;
 let latestAuthSuccessAt = 0;
 
-function markUserAuthenticated(): void {
+export function markUserAuthenticated(): void {
   userWasAuthenticated = true;
   latestAuthSuccessAt =
     typeof performance !== "undefined" ? performance.now() : Date.now();
@@ -1450,174 +1450,19 @@ export {
   type CreatedApiKey,
 } from "@/api/user-management-api";
 
-// ALERTS
-// ============================================================================
+export {
+  setupTOTP,
+  enableTOTP,
+  disableTOTP,
+  verifyTOTPLogin,
+  generateBackupCodes,
+  getUserAlerts,
+  dismissAlert,
+  getReleasesRSS,
+  getVersionInfo,
+  getDatabaseHealth,
+} from "@/api/system-status-api";
 
-export async function setupTOTP(): Promise<{
-  secret: string;
-  qr_code: string;
-}> {
-  try {
-    const response = await authApi.post("/users/totp/setup");
-    return response.data;
-  } catch (error) {
-    handleApiError(error as AxiosError, "setup TOTP");
-    throw error;
-  }
-}
-
-export async function enableTOTP(
-  totp_code: string,
-): Promise<{ message: string; backup_codes: string[] }> {
-  try {
-    const response = await authApi.post("/users/totp/enable", { totp_code });
-    return response.data;
-  } catch (error) {
-    handleApiError(error as AxiosError, "enable TOTP");
-    throw error;
-  }
-}
-
-export async function disableTOTP(
-  password?: string,
-  totp_code?: string,
-): Promise<{ message: string }> {
-  try {
-    const response = await authApi.post("/users/totp/disable", {
-      password,
-      totp_code,
-    });
-    return response.data;
-  } catch (error) {
-    handleApiError(error as AxiosError, "disable TOTP");
-    throw error;
-  }
-}
-
-export async function verifyTOTPLogin(
-  temp_token: string,
-  totp_code: string,
-  rememberMe: boolean = false,
-): Promise<AuthResponse> {
-  try {
-    const response = await authApi.post("/users/totp/verify-login", {
-      temp_token,
-      totp_code,
-      rememberMe,
-    });
-
-    const isInIframe =
-      typeof window !== "undefined" && window.self !== window.top;
-
-    if (isInIframe && isElectron() && response.data.success) {
-      try {
-        window.parent.postMessage(
-          {
-            type: "AUTH_SUCCESS",
-            source: "totp_verify",
-            platform: "desktop",
-            timestamp: Date.now(),
-          },
-          window.location.origin,
-        );
-      } catch (e) {
-        console.error("[main-axios] Error posting message to parent:", e);
-      }
-    }
-
-    if (response.data.success) {
-      markUserAuthenticated();
-    }
-
-    return response.data;
-  } catch (error) {
-    handleApiError(error as AxiosError, "verify TOTP login");
-    throw error;
-  }
-}
-
-export async function generateBackupCodes(
-  password?: string,
-  totp_code?: string,
-): Promise<{ backup_codes: string[] }> {
-  try {
-    const response = await authApi.post("/users/totp/backup-codes", {
-      password,
-      totp_code,
-    });
-    return response.data;
-  } catch (error) {
-    handleApiError(error as AxiosError, "generate backup codes");
-    throw error;
-  }
-}
-
-export async function getUserAlerts(): Promise<{
-  alerts: Array<Record<string, unknown>>;
-}> {
-  try {
-    const response = await authApi.get(`/alerts`);
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "fetch user alerts");
-    throw error;
-  }
-}
-
-export async function dismissAlert(
-  alertId: string,
-): Promise<Record<string, unknown>> {
-  try {
-    const response = await authApi.post("/alerts/dismiss", { alertId });
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "dismiss alert");
-    throw error;
-  }
-}
-
-// ============================================================================
-// UPDATES & RELEASES
-// ============================================================================
-
-export async function getReleasesRSS(
-  perPage: number = 100,
-): Promise<Record<string, unknown>> {
-  try {
-    const response = await authApi.get(`/releases/rss?per_page=${perPage}`);
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "fetch releases RSS");
-  }
-}
-
-export async function getVersionInfo(
-  checkRemote = true,
-): Promise<Record<string, unknown>> {
-  try {
-    const response = await authApi.get(
-      `/version${checkRemote ? "" : "?checkRemote=false"}`,
-    );
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "fetch version info");
-  }
-}
-
-// ============================================================================
-// DATABASE HEALTH
-// ============================================================================
-
-export async function getDatabaseHealth(): Promise<Record<string, unknown>> {
-  try {
-    const response = await authApi.get("/health");
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "check database health");
-  }
-}
-
-// ============================================================================
 // SSH CREDENTIALS MANAGEMENT
 // ============================================================================
 
@@ -1674,87 +1519,13 @@ export {
 } from "@/api/snippets-api";
 
 // ============================================================================
-// DASHBOARD API
-// ============================================================================
-
-export interface UptimeInfo {
-  uptimeMs: number;
-  uptimeSeconds: number;
-  formatted: string;
-}
-
-export interface RecentActivityItem {
-  id: number;
-  userId: string;
-  type:
-    | "terminal"
-    | "file_manager"
-    | "server_stats"
-    | "tunnel"
-    | "docker"
-    | "telnet"
-    | "vnc"
-    | "rdp";
-  hostId: number;
-  hostName: string;
-  timestamp: string;
-}
-
-export async function getUptime(): Promise<UptimeInfo> {
-  try {
-    const response = await dashboardApi.get("/uptime");
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch uptime");
-  }
-}
-
-export async function getRecentActivity(
-  limit?: number,
-): Promise<RecentActivityItem[]> {
-  try {
-    const response = await dashboardApi.get("/activity/recent", {
-      params: { limit },
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch recent activity");
-  }
-}
-
-export async function logActivity(
-  type:
-    | "terminal"
-    | "file_manager"
-    | "server_stats"
-    | "tunnel"
-    | "docker"
-    | "rdp"
-    | "vnc"
-    | "telnet",
-  hostId: number,
-  hostName: string,
-): Promise<{ message: string; id: number | string }> {
-  try {
-    const response = await dashboardApi.post("/activity/log", {
-      type,
-      hostId,
-      hostName,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "log activity");
-  }
-}
-
-export async function resetRecentActivity(): Promise<{ message: string }> {
-  try {
-    const response = await dashboardApi.delete("/activity/reset");
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "reset recent activity");
-  }
-}
+export type { UptimeInfo, RecentActivityItem } from "@/api/dashboard-api";
+export {
+  getUptime,
+  getRecentActivity,
+  logActivity,
+  resetRecentActivity,
+} from "@/api/dashboard-api";
 
 // ============================================================================
 export {
